fix(img): Add jpg as permitted image type #58
Default image types didnt allow for jpg. Added and also updated the error
message that is fed back to the user in the event that a file they have
attempted to upload has the correct extension type, but is not the
correct file type. e.g. a PDF file has had its extension changed to jpg.

closes #58