fix(toolbar): Refactor and add check for options #63
Renamed the toolbar options config to make it clear what the purpose
is.
Removed the hard coded default toolbar config from the widget and use
the settings default toolbar.
Added a check in the widget so that if the toolbar config is not in
the list of options the default toolbar will be selected.

closes #63